requirements.txt
# Learning Automata Kelp Detection Prototype
# Requirements for the kelp conda environment (Python 3.10)

# Core scientific computing
numpy>=1.23.0
scipy>=1.9.0

# Deep learning framework
torch>=2.0.0

# Geospatial data handling
rasterio>=1.3.0
GDAL>=3.4.0

# Data processing
pandas>=1.5.0
scikit-image>=0.19.0

# Configuration and experiment tracking
pyyaml>=6.0
tqdm>=4.64.0

# Visualization
matplotlib>=3.5.0

# Testing
pytest>=7.0.0
